requirements.txt
streamlit
numpy
pandas
scikit-learn
requests
openai
mistralai
sentence-transformers
faiss-cpu